fix: Resolve TypeScript compiler errors in DTO and game store
- Extract DtoStringToDate/DtoStringToDuration to dto-types.ts to break circular dependency between dto.ts and model files
- Remove spurious .value access on Pinia auto-unwrapped computed ref in featureGameStore
- Simplify convertTimerDto by removing unnecessary undefined guard on non-optional field
