+++ src/api-facade/models/games-models.ts
@@ -1,4 +1,4 @@
-import { type DtoStringToDate, type DtoStringToDuration } from '../dto'
+import { type DtoStringToDate, type DtoStringToDuration } from '../dto-types'
 
 export type WishlistedGame = {
 	name: string
